+++ packages/devextreme/js/__internal/grids/data_grid/module_not_extended/virtual_scrolling.ts
@@ -1,4 +1,5 @@
-import { dataSourceAdapterExtender, virtualScrollingModule } from '@ts/grids/grid_core/virtual_scrolling/m_virtual_scrolling';
+import { dataSourceAdapterExtender } from '@ts/grids/grid_core/virtual_scrolling/m_virtual_scrolling';
+import { virtualScrollingModule } from '@ts/grids/grid_core/virtual_scrolling/virtual_scrolling_module';
 
 import gridCore from '../m_core';
 import dataSourceAdapterProvider from '../m_data_source_adapter';
