+++ packages/devextreme/js/__internal/grids/grid_core/virtual_scrolling/const.ts
@@ -1,0 +1,17 @@
+export const BOTTOM_LOAD_PANEL_CLASS = 'bottom-load-panel';
+export const GROUP_SPACE_CLASS = 'group-space';
+export const FREESPACE_CLASS = 'dx-freespace-row';
+export const COLUMN_LINES_CLASS = 'dx-column-lines';
+export const VIRTUAL_ROW_CLASS = 'dx-virtual-row';
+export const ROW_INSERTED = 'dx-row-inserted';
+
+export const SCROLLING_MODE_INFINITE = 'infinite';
+export const SCROLLING_MODE_VIRTUAL = 'virtual';
+export const LOAD_TIMEOUT = 300;
+export const LEGACY_SCROLLING_MODE = 'scrolling.legacyMode';
+export const VISIBLE_PAGE_INDEX = 'paging.pageIndex';
+export const PAGING_METHOD_NAMES = [
+  'beginPageIndex',
+  'endPageIndex',
+  'pageIndex',
+];
